fix: move debug functions to other file
