+++ app/routes/admin/debug.tsx
@@ -1,11 +1,11 @@
 import { ActionFunction } from "@remix-run/node";
 import { useSubmit } from "@remix-run/react";
 import { useRef } from "react";
-import { debug_fixPartnerProviderNameStandard } from "~/services/firebase.server";
+import { debug_initializeIsDiscountedForSettlements } from "~/services/firebase-debug.server";
 
 export const action: ActionFunction = async ({ request }) => {
   console.log("request submitted");
-  debug_fixPartnerProviderNameStandard();
+  debug_initializeIsDiscountedForSettlements();
   return null;
 };
