+++ app/services/firebase.server.ts
@@ -1943,110 +1943,7 @@ export async function sendSettlementNoticeEmail({
   };
 }
 
-export async function debug_fixProductStorage() {
-  const productsRef = collection(firestore, `products`);
-  const productsQuery = query(
-    productsRef,
-    where("partnerName", "==", "homethus")
-  );
-  const querySnapshot = await getDocs(productsQuery);
-  const usages = ["main", "thumbnail", "detail", "extra"];
-  querySnapshot.forEach(async (item) => {
-    const docId = item.id;
-    const data = item.data();
-    const id = data.id;
-    const updateDataJson: any = {
-      detailImageURLList: [],
-      extraImageURLList: [],
-    };
 
-    // 폴더 내의 모든 파일과 하위 폴더 목록 가져오기
-    const promises = usages.map(async (usage) => {
-      const oldFolderRef = ref(storage, `${data.productName}/${usage}`);
-      const listResult = await listAll(oldFolderRef);
-      console.log(
-        `${data.productName}/${usage} length`,
-        listResult.items.length
-      );
-
-      // 모든 파일 복사 및 삭제
-      const promises = listResult.items.map(async (itemRef, index) => {
-        console.log("PROMISE");
-        const newFolderRef = ref(
-          storage,
-          `product-images/${id}/${usage}/${usage}_${index}`
-        );
-        const downloadURL = await getDownloadURL(itemRef);
-        console.log(
-          `${data.productName}/${usage} ${index} previous downloadURL`,
-          downloadURL
-        );
-        const response = await fetch(downloadURL);
-        const fileData = await response.blob(); // Get file data as a Blob
-        await uploadBytes(newFolderRef, fileData).then(async (snapshot) => {
-          const url = await getDownloadURL(snapshot.ref);
-          console.log(
-            `product-images/${id}/${usage}/${usage}_${index} new downloadURL`,
-            url
-          );
-          switch (usage) {
-            case "main":
-              updateDataJson["mainImageURL"] = url;
-              break;
-            case "thumbnail":
-              updateDataJson["thumbnailImageURL"] = url;
-              break;
-            case "detail":
-              updateDataJson["detailImageURLList"].push(url);
-              break;
-            case "extra":
-              updateDataJson["extraImageURLList"].push(url);
-              break;
-          }
-        });
-        return deleteObject(itemRef);
-      });
-
-      await Promise.all(promises);
-
-      // 모든 삭제 작업이 완료될 때까지 기다림
-      console.log(`All files in folder '${id} ${usage}' have been rename.`);
-    });
-    await Promise.all(promises);
-    console.log("update", updateDataJson);
-    await updateDoc(doc(firestore, "products", docId), updateDataJson);
-  });
-}
-
-export async function debug_fixPartnerProfileTaxStandard() {
-  const accountsRef = collection(firestore, "accounts");
-  const querySnap = await getDocs(accountsRef);
-  querySnap.docs.forEach(async (item) => {
-    const data = item.data();
-    if (!data.businessTaxStandard) {
-      await updateDoc(doc(firestore, "accounts", data.name), {
-        businessTaxStandard: "일반",
-      }).catch((error) => {
-        return error.message;
-      });
-    }
-  });
-}
-
-export async function debug_fixPartnerProviderNameStandard() {
-  const accountsRef = collection(firestore, "accounts");
-  const querySnap = await getDocs(accountsRef);
-  querySnap.docs.forEach(async (item) => {
-    const data = item.data();
-    if (!data.providerName) {
-      await updateDoc(doc(firestore, "accounts", data.name), {
-        providerName: data.name,
-      }).catch((error) => {
-        return error.message;
-      });
-    }
-  });
-}
 
 /**
  * 수익통계자료를 업로드합니다.
